updated lib testers to be separate classes
