+++ tests/jenkins/lib-testers/TestUploadToS3LibTester.groovy
@@ -1,0 +1,42 @@
+import static org.hamcrest.CoreMatchers.notNullValue
+import static org.hamcrest.MatcherAssert.assertThat
+
+class TestUploadToS3LibTester extends LibFunctionTester {
+
+    private String sourcePath
+    private String bucket
+    private String path
+
+    public TestUploadToS3LibTester(sourcePath, bucket, path){
+        this.sourcePath = sourcePath
+        this.bucket = bucket
+        this.path = path
+    }
+
+    void parameterInvariantsAssertions(call){
+        assertThat(call.args.sourcePath.first(), notNullValue())
+        assertThat(call.args.bucket.first(), notNullValue())
+        assertThat(call.args.path.first(), notNullValue())
+    }
+
+    boolean expectedParametersMatcher(call) {
+        return call.args.sourcePath.first().toString().equals(this.sourcePath)
+                && call.args.bucket.first().toString().equals(this.bucket)
+                && call.args.path.first().toString().equals(this.path)
+    }
+
+    String libFunctionName(){
+        return 'uploadToS3'
+    }
+
+    void configure(helper, binding){
+        binding.setVariable('ARTIFACT_UPLOAD_ROLE_NAME', 'Dummy_Upload_Role')
+        binding.setVariable('AWS_ACCOUNT_PUBLIC', 'dummy_account')
+        binding.setVariable('ARTIFACT_BUCKET_NAME', 'dummy_bucket_name')
+        helper.registerAllowedMethod("s3Upload", [Map])
+        helper.registerAllowedMethod("withAWS", [Map, Closure], { args, closure ->
+            closure.delegate = delegate
+            return helper.callClosure(closure)
+        })
+    }
+}
